sql: support expressions in unique constraints
Postgres does not allow expressions in unique constraints, but it does
allow expressions in unique indexes. For now we allow expressions in
unique constraints, because we cannot differentiate between a unique
constraint and a unique index table definition: they are both parsed
into the same struct, `tree.UniqueConstraintTableDef`.

In the long term we may want to disallow expression in unique
constraints to be consistent with Postgres. We could do this by changing
the parser so that `UNIQUE ((a + b))` does not parse successfully.
See #65825.

Release note: None