roachtest: allow 10 second buffer before scraping toTime for DRT
The scrape we do after before the preShutdown or preStartup can be
slightly racey and may pick up an extra error or success when a node has
already begun to shutdown or restart. This commit adjusts for that by
scraping 10 seconds before the pre event.

Also increase the warehouses because we can now.

Release note: None