democluster: pull the interactive part of the API into separate pkg
In the longer term, we will want the ability to instantiate a SQL
interactive shell without the need to link in all the server code.

This requires the ability to compile the SQL shell without a
dependency on the `democluster` package.

Towards this goal, this commit extracts the part of the
`democluster.DemoCluster` interface API used in the SQL shell into a
separate package.

Release note: None